Remove old monaco editor code & packages
