+++ .gitignore
@@ -23,3 +23,4 @@ yarn-error.log*
 _test_expressions.txt
 _test_number_grouping.txt
 _test_ace_mode.txt
+_test_ace_mode_expressions.txt
