+++ .prettierignore
@@ -1,3 +1,2 @@
-src/monaco/calcLang.ts
 public/css/_fonts.scss
 src/css/_fonts.scss
