+++ package.json
@@ -18,7 +18,6 @@
     "css-loader": "^5.2.4",
     "html-webpack-plugin": "^5.3.1",
     "mini-css-extract-plugin": "^1.6.0",
-    "monaco-editor-webpack-plugin": "^3.0.1",
     "prettier": "^2.2.1",
     "sass": "^1.32.12",
     "sass-loader": "^11.0.1",
@@ -30,7 +29,6 @@
   },
   "dependencies": {
     "ace-builds": "^1.4.12",
-    "mathjs": "^9.3.2",
-    "monaco-editor": "^0.23.0"
+    "mathjs": "^9.3.2"
   }
 }
