+++ src/css/app.scss
@@ -89,15 +89,6 @@ a:hover {
       overflow-y: hidden;
     }
 
-    /* textarea {
-      line-height: var(--input-lh);
-      border: none;
-      background-color: transparent;
-      color: var(--input-font-color);
-      outline: none;
-      resize: none;
-    } */
-
     .mathResult {
       color: var(--result-font-color);
